divided code to Bridge and Bulb
added new license, probably should look over everything again and add
the little GPLv3 header thing to anywhere it isn't already.

bridge code needs cleanup, it's a bit messy and has lots of commented
code copied from examples.  multicast would be nice.

changed BROKER_IP in Bridge node back to ad-hoc IP, as base station
is back in working order, for the time being, but that isn't in this
repo.

can remove Private.h from LedReactorBulb, but LedReactorBridge either
needs a blank generic one, or needs macros pasted in main.cpp.

added library.json, because it seemed like the right thing to do.

added LedWriter as a dependency, but the library itself is not included.
suppose I should figure how to upload directly to PlatformIO to use it
as one, else the code is located at:

https://github.com/khaudio/LedWriter